### 2025-11-18 #9463 @editorialbot
[PRE REVIEW]: SOFIA: A Python Library for High-Quality 2D Triangular Mesh Adaptation
**Submitting author:** @youssef-mesri (<a href="http://orcid.org/0000-0002-5136-5435">Youssef Mesri</a>)
**Repository:** https://github.com/youssef-mesri/sofia-mesh
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
