### 2025-11-22 fab3a7d
Update README to remove old documentation links
Removed outdated sections and links from the README.
