+++ docs/README.md
@@ -7,19 +7,10 @@ This directory contains all project documentation.
 - **[CITATION.md](CITATION.md)** - How to cite SOFIA in your work
 - **[CODE_OF_CONDUCT.md](CODE_OF_CONDUCT.md)** - Community guidelines
 - **[CONTRIBUTING.md](CONTRIBUTING.md)** - How to contribute to SOFIA
-- **[PUBLICATION_GUIDE.md](PUBLICATION_GUIDE.md)** - Complete guide for publishing SOFIA
-- **[QUICK_START.md](QUICK_START.md)** - Quick start guide for publication
 
 ## Images
 
 The `images/` directory contains all documentation images and demos.
-
-## Developer Documentation
-
-See the `dev/` directory for:
-- **[MOVING_MODULES.md](dev/MOVING_MODULES.md)** - Module organization
-- **[REFACTORING_VISUAL.txt](dev/REFACTORING_VISUAL.txt)** - Refactoring notes
-- **archive/** - Historical preparation documents
 
 ## Additional Documentation
 
