### 2025-11-25 2a348a0
fix joss paper references format and reduce the wordcount to fit the journal requirement
