+++ paper.bib
@@ -42,7 +42,8 @@
   volume={4},
   pages={69--190},
   year={1996},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/s1570-8659(96)80003-2}
 }
 
 @article{Borouchaki1998,
@@ -72,6 +73,8 @@
   author={Kloeckner, Andreas},
   title={MeshPy: Simplicial Mesh Generation from Python},
   url={https://mathema.tician.de/software/meshpy/},
+  publisher = {Mathema},
+  journal = {Website repository},
   year={2015}
 }
 
@@ -79,6 +82,8 @@
   author={Zhou, Qingnan},
   title={PyMesh: Geometry Processing Library for Python},
   url={https://github.com/PyMesh/PyMesh},
+  publisher = {GitHub},
+  journal = {GitHub repository},
   year={2018}
 }
 
@@ -135,7 +140,7 @@
   pages={247--258},
   year={2010},
   publisher={Springer},
-  doi={10.1007/s00366-009-0146-3}
+  doi={10.1007/s00366-005-0009-3}
 }
 
 @article{Loseille2011,
@@ -154,10 +159,11 @@
 @article{mesri2006continuous,
   title={Continuous mesh adaptation models for CFD},
   author={Mesri, Youssef and Alauzet, Fr{\'e}d{\'e}ric and Loseille, Adrien and Hasco{\"e}t, Laurent and Koobus, Bruno and Dervieux, Alain},
-  journal={CFD Journal},
+  journal={Computational fluid dynamics journal},
   volume={16},
   number={4},
-  year={2006}
+  pages={346--355},
+  year={2008}
 }
 
 @incollection{mesri2008dynamic,
