+++ paper.md
@@ -20,23 +20,23 @@ bibliography: paper.bib
 
 # Summary
 
-Triangular meshes are fundamental data structures in computational science and engineering, serving as the backbone for numerical simulations in fluid dynamics, structural analysis, heat transfer, and many other fields. The quality of a mesh directly impacts the accuracy, stability, and convergence of numerical solutions \cite{Shewchuk1996,Geuzaine2009,Persson2004}. `SOFIA` (Scalable Operators for Field-driven Iso/Ani Adaptation) is a Python library that provides robust and efficient tools for 2D triangular mesh modification, refinement, and quality improvement through local topological operations, with particular emphasis on **anisotropic mesh adaptation** and **automatic boundary preservation**.
+Triangular meshes are fundamental data structures in computational science and engineering, serving as the backbone for numerical simulations in fluid dynamics, structural analysis, heat transfer, and many other fields. The quality of a mesh directly impacts the accuracy, stability, and convergence of numerical solutions [@mesri2008dynamic;@manzinali2018adaptive;@Persson2004]. `SOFIA` (Scalable Operators for Field-driven Iso/Ani Adaptation) is a Python library that provides robust and efficient tools for 2D triangular mesh modification, refinement, and quality improvement through local topological operations, with particular emphasis on **anisotropic mesh adaptation** and **automatic boundary preservation**.
 
 # Statement of Need
 
-While several mesh generation tools exist \cite{Shewchuk1996,Geuzaine2009,MeshPy,PyMesh}, there is a gap in the Python ecosystem for a lightweight, well-documented library focused specifically on **mesh adaptation** and **local modification operations**. Existing solutions either require complex dependencies (C/C++ bindings), lack comprehensive documentation, or focus primarily on initial mesh generation rather than adaptive refinement during simulation workflows.
+While several mesh generation tools exist [@Shewchuk1996;@Geuzaine2009;@MeshPy;@PyMesh], there is a gap in the Python ecosystem for a lightweight, well-documented library focused specifically on **mesh adaptation** and **local modification operations**. Existing solutions either require complex dependencies (C/C++ bindings), lack comprehensive documentation, or focus primarily on initial mesh generation rather than adaptive remeshing during simulation workflows.
 
 `SOFIA` addresses these needs by providing:
 
 1. **Pure Python implementation with optional C++ acceleration**: Easy to install and modify, with optional C++ backend for performance-critical operations
-2. **Local topological operations**: Fine-grained control over mesh modification through fundamental operations (split, collapse, flip, insert, remove)
-3. **Anisotropic mesh adaptation**: Metric-based adaptation with automatic handling of highly elongated elements near features (boundary layers, shocks, fronts)
-4. **Automatic boundary preservation**: Smart edge collapse that automatically detects and preserves domain boundaries without manual vertex protection
-5. **Quality-driven adaptation**: Built-in quality metrics and optimization strategies, with separate handling for isotropic/anisotropic meshes
+2. **Local topological operations**: Fine-grained control over mesh modification through fundamental operations (split, collapse, flip, add, remove)
+3. **Anisotropic mesh adaptation**: Metric-based adaptation with automatic handling of highly stretched elements near features (boundary layers, shocks, fronts)
+4. **Automatic boundary preservation**: Edge collapse that automatically detects and preserves domain boundaries
+5. **Quality-driven adaptation**: Built-in quality metrics and optimization strategies handling both isotropic/anisotropic meshes
 6. **Boundary-aware operations**: Safe manipulation of mesh boundaries with conformity preservation
[…]
