### 2025-11-25 7936cea
delete preparation docs for joss submission
