+++ docs/PUBLICATION_GUIDE.md
@@ -1,414 +1,0 @@
-# SOFIA - Guide Complet de Publication
-
-**Date de préparation:** 24 octobre 2025  
-**Statut:** PRÊT POUR PUBLICATION
-
----
-
-## Vue d'ensemble
-
-Ce guide vous accompagne étape par étape pour publier SOFIA sur GitHub et PyPI.
-
-### Ce qui est prêt
-
-**8 exemples complets** (2144 lignes de code)  
-**Documentation complète** (README_NEW.md, 230 lignes)  
-**Tests automatisés** (GitHub Actions configuré)  
-**Citations académiques** (CITATION.cff + CITATION.md)  
-**Fichiers de packaging** (pyproject.toml, setup.py)  
-**Visualisations** (8 PNG, 2.4 MB)  
-**Code of Conduct** (CODE_OF_CONDUCT.md)  
-**Guide de contribution** (CONTRIBUTING.md)
-
----
-
-## Checklist Avant Publication
-
-### 1. Vérifications Finales
-
-- [ ] **Mettre à jour votre ORCID** dans README_NEW.md (ligne ~219)
-  ```bash
-  # Éditer et remplacer 0000-0002-XXXX-XXXX par votre vrai ORCID
-  nano README_NEW.md
-  ```
-
-- [ ] **Vérifier l'email** dans pyproject.toml
-  ```bash
-  grep "email" pyproject.toml
-  # Devrait afficher: youssef.mesri@minesparis.psl.eu
-  ```
-
-- [ ] **Tester tous les exemples**
-  ```bash
-  cd /home/ymesri/Sofia/publication_prep
-  for example in examples/*.py; do
-    echo "Testing $example..."
-    python "$example" || echo " Failed: $example"
-  done
-  ```
-
-- [ ] **Lancer les tests unitaires**
-  ```bash
-  pytest tests/ -v
-  ```
-
-### 2. Remplacer le README
-
-```bash
-cd /home/ymesri/Sofia/publication_prep
-
-# Sauvegarder l'ancien README
-mv README.md README_OLD.md
-
-# Activer le nouveau README
-cp README_NEW.md README.md
-
-# Vérifier
-head -20 README.md
-```
-
-### 3. Nettoyer les fichiers temporaires
-
-```bash
-# Supprimer les fichiers de préparation (optionnel)
-rm -f PREPARATION_STATUS.md PUBLICATION_COMPLETE.md PUBLICATION_GUIDE.md
-
-# Supprimer les anciens benchmarks (optionnel)
-rm -f benchmark_*.py batch_benchmark_results.json phase2_results.json
-
-# Supprimer les fichiers RL (optionnel)
-rm -f rl-ym.py remesh_*.py smart_ppo_*.py test_remesh_env.py
-
-# Garder seulement les PNG des exemples dans examples/
-mkdir -p examples/visualizations
-mv *_result.png examples/visualizations/ 2>/dev/null || true
-```
-
----
-
-## Publication sur GitHub
-
-### Étape 1: Créer le dépôt public
-
-1. **Aller sur GitHub:** https://github.com/new
-
-2. **Paramètres du dépôt:**
-   - **Repository name:** `sofia`
-   - **Description:** "SOFIA - Scalable Operators for Field-driven Iso/Ani Adaptation: A Modern 2D Triangular Mesh Modification Library"
-   - **Visibility:** Public
-   - **Initialize:** Ne pas initialiser (déjà fait localement)
-
-3. **Cliquer sur "Create repository"**
-
-### Étape 2: Pousser le code
-
-```bash
-cd /home/ymesri/Sofia/publication_prep
-
-# Ajouter le remote (si pas déjà fait)
-git remote add public https://github.com/youssef-mesri/sofia.git
-
-# Vérifier les remotes
[…]
