+++ docs/QUICK_START.md
@@ -1,70 +1,0 @@
-# Quick Start - Publication de SOFIA
-
-## Statut: PRÊT
-
-### Pour Publier (1 heure)
-
-```bash
-cd /home/ymesri/Sofia/publication_prep
-
-# 1. Vérification finale
-python verify_publication.py
-
-# 2. Tester tous les exemples
-./test_examples.sh
-
-# 3. Activer le nouveau README (optionnel)
-mv README.md README_OLD.md && cp README_NEW.md README.md
-
-# 4. Build & test sur TestPyPI
-pip install --upgrade build twine
-python -m build
-twine upload --repository testpypi dist/*
-
-# 5. Publier sur GitHub
-git add -A
-git commit -m "feat: Prepare for public release v0.1.0"
-git push origin main
-
-# 6. Créer release sur GitHub
-# https://github.com/youssef-mesri/sofia/releases/new
-
-# 7. Publier sur PyPI
-twine upload dist/*
-```
-
-### Fichiers Clés
-
-- **PUBLICATION_GUIDE.md** - Guide détaillé complet
-- **SESSION_SUMMARY.md** - Résumé de cette session
-- **verify_publication.py** - Script de vérification
-- **test_examples.sh** - Test rapide des exemples
-
-### Ce qui est Prêt
-
-11 exemples complets (2144 lignes)  
-8 visualisations (2.4 MB)  
-Documentation complète (1059 lignes)  
-Tests unitaires (50+)  
-GitHub Actions (CI/CD)  
-Package PyPI configuré  
-Citations académiques  
-Code de conduite
-
-### Nouveautés
-
-**boundary_refinement.py** - Raffinement du bord (domaine circulaire)  
-**combined_refinement.py** - Raffinement multi-critères (domaine en L)
-
-### Support
-
-Questions? Voir **PUBLICATION_GUIDE.md** sections:
-- Vérifications finales
-- Test sur TestPyPI
-- Publication GitHub
-- Publication PyPI
-- Dépannage
-
----
-
-**Temps estimé:** ~1h pour publication complète
