### 2025-11-25 07e6aae
reduce the wordcount to fit the journal requirement
