### 2025-11-26 5a9c070
omit API.md from docs
