### 2026-01-05 #9729 @editorialbot
[REVIEW]: SOFIA: A Python Library for High-Quality 2D Triangular Mesh Adaptation
**Submitting author:** @youssef-mesri (<a href="http://orcid.org/0000-0002-5136-5435">Youssef Mesri</a>)
**Repository:** https://github.com/youssef-mesri/sofia-mesh
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
