### 2026-01-14 #9729 @danielskatz
👋 @inducer - @arfon can update your checklist to our old version. Because this submission came in before JOSS changed a few things recently, it's being treated under the old guidelines, but our system is generating the new checklist.  The development period wasn't part of our old checklist...
