### 2026-01-26 youssef-mesri/sofia-mesh#3 @joewallwork
[JOSS review]: `CONTRIBUTING.md` doesn't exist
Towards https://github.com/openjournals/joss-reviews/issues/9729

The README links to https://github.com/youssef-mesri/sofia-mesh/blob/main/CONTRIBUTING.md but this file doesn't exist.
