### 2026-01-26 eed929e
Fix path to `CONTRIBUTING.md`
Closes #3.
