+++ README.md
@@ -134,7 +134,7 @@ print(f"Final quality: {improved_editor.mesh_min_angle():.2f}°")
 
 - **[API Reference](docs/API_REFERENCE.md)** - Complete API documentation
 - **[Examples](examples/)** - Practical usage examples
-- **[Contributing](CONTRIBUTING.md)** - How to contribute
+- **[Contributing](docs/CONTRIBUTING.md)** - How to contribute
 - **[Publication Strategy](docs/PUBLICATION_STRATEGY.md)** - Development roadmap
 
 ---
