### 2026-01-26 youssef-mesri/sofia-mesh#5 @joewallwork
[JOSS review] Broken link to API reference docs
Towards https://github.com/openjournals/joss-reviews/issues/9729

The link to the API reference docs on the following line doesn't work
https://github.com/youssef-mesri/sofia-mesh/blob/5a9c07027358d114213e19b0ca6da90427b533d4/README.md?plain=1#L135

Looking at https://github.com/youssef-mesri/sofia-mesh/tree/main/docs, I can't see this file. Where do the API reference docs exist?
