### 2026-01-26 youssef-mesri/sofia-mesh#6 @joewallwork
[JOSS review] Version inconsistency
Towards https://github.com/openjournals/joss-reviews/issues/9729

The latest tagged version in the repo is v0.1.0. However, the README mentions that the current version is v1.0. Is there a missing tagged release or is the README incorrect?
