### 2026-01-27 #9729 @joewallwork
Thank you for this submission, SOFIA seems like a really nice tool to have available as pure Python. I haven't had chance to experiment with it yet and will do the rest of the software review separately, but here are my initial thoughts on the paper:

* General feedback
	* I notice paper doesn't refer to the mesh movement literature. The listed core operations also don't mention node movement. Is Sofia purely $h$-adaptive, or can it do $r$-adaptation, too? (There is mention of vertex smoothing, but I imagine this is to boost the mesh quality and align with boundaries following more significant $h$-adaptation.) Also, is there any distinction between metric-based and hierarchical $h$-adaptation methods that would be worth making? I feel this kind of information should be clarified in the paper and perhaps even the title.
	* Is there perhaps a Figure that you could include in the paper? For example, an image of an example of an anisotropic adapted mesh generated with SOFIA that shows off some of the stated highlights would help the reader to understand what's being achieved.
	* The paper contains many bullet point lists and enumerations. I am happy for bullet points to be used but I feel the paper may read better if some of the lists were converted into paragraphs of text. For example, the "key innovation" section could be converted into a paragraph.
* Statement of need section
	* C++ acceleration is mentioned but the repo README says this will come with the future v2.0 release. I notice that the most recent tag is v0.1. I don't think the C++ acceleration can be mentioned in the main part of the paper (i.e., as something other than future work) if it isn't something that's currently available.
* Anisotropic mesh adaptation section
	* "up to 10:1 or higher" seems a bit redundant. Why is the 10:1 ratio worth mentioning here?
	* Need to clarify the maths notation used because it's not defined anywhere else.
* State of the field section
	* Recommend including the citations in the bullet points so it's clearer which packages they correspond to.
	* The "Key differentiators" part seems to be an incorrectly formatted bullet point list. Again, I'm not sure it really needs to be bullets, though.
	* You mention Firedrake mesh tools but don't provide a reference. [Animate](https://github.com/mesh-adaptation/animate) provides an interface for metric-based mesh adaptation in Firedrake. My collaborators and I are currently working on a JOSS submission for this, but unfortunately it isn't ready just yet. However, we would appreciate if you could cite the repo and/or relevant paper for coupling [Par]Mmg into PETSc (which we pipe through to Firedrake): Wallwork et al., Parallel metric-based mesh adaptation in PETSc using ParMmg, in: SIAM International Meshing Roundtable, 2022. arxiv.org/abs/2201.02806.
* References
	* Ensure you use consistent capitalisation in the references e.g., Python, part I, Hessian, Delaunay
