### 2026-01-29 a09cde3
Update JOSS paper per reviewer feedback
