+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Shewchuk1996,
-  title={Triangle: Engineering a 2D quality mesh generator and Delaunay triangulator},
+  title={Triangle: Engineering a 2D quality mesh generator and {Delaunay} triangulator},
   author={Shewchuk, Jonathan Richard},
   journal={Applied computational geometry towards geometric engineering},
   pages={203--222},
@@ -9,7 +9,7 @@
 }
 
 @article{Ruppert1995,
-  title={A Delaunay refinement algorithm for quality 2-dimensional mesh generation},
+  title={A {Delaunay} refinement algorithm for quality 2-dimensional mesh generation},
   author={Ruppert, Jim},
   journal={Journal of algorithms},
   volume={18},
@@ -21,7 +21,7 @@
 }
 
 @article{Cheng2002,
-  title={Delaunay mesh generation},
+  title={{Delaunay} mesh generation},
   author={Cheng, Siu-Wing and Dey, Tamal K and Shewchuk, Jonathan},
   year={2012},
   publisher={CRC Press},
@@ -47,7 +47,7 @@
 }
 
 @article{Borouchaki1998,
-  title={Delaunay mesh generation governed by metric specifications. Part I. Algorithms},
+  title={{Delaunay} mesh generation governed by metric specifications. {Part I}. Algorithms},
   author={Borouchaki, Houman and Hecht, Fr{\'e}d{\'e}ric and Frey, Pascal J},
   journal={Finite elements in analysis and design},
   volume={25},
@@ -87,6 +87,22 @@
   year={2018}
 }
 
+@software{MMG,
+  author    = {{MmgTools contributors}},
+  title     = {{Mmg}: Open source software for simplicial mesh modification},
+  url       = {https://github.com/MmgTools/mmg},
+  publisher = {GitHub},
+  journal   = {GitHub repository}
+}
+
+@software{BAMG,
+  author    = {{FreeFEM contributors}},
+  title     = {{BAMG}: Bidimensional anisotropic mesh generator/adaptation},
+  url       = {https://github.com/FreeFem/FreeFem-sources},
+  publisher = {GitHub},
+  journal   = {GitHub repository}
+}
+
 @article{Logg2012,
   title={Automated solution of differential equations by the finite element method},
   author={Logg, Anders and Mardal, Kent-Andre and Wells, Garth N},
@@ -108,7 +124,7 @@
 }
 
 @article{Persson2004,
-  title={A simple mesh generator in MATLAB},
+  title={A simple mesh generator in {MATLAB}},
   author={Persson, Per-Olof and Strang, Gilbert},
   journal={SIAM review},
   volume={46},
@@ -144,7 +160,7 @@
 }
 
 @article{Loseille2011,
-  title={Continuous mesh framework part I: well-posed continuous interpolation error},
+  title={Continuous mesh framework {Part I}: well-posed continuous interpolation error},
   author={Loseille, Adrien and Alauzet, Fr{\'e}d{\'e}ric},
   journal={SIAM Journal on Numerical Analysis},
   volume={49},
@@ -178,7 +194,7 @@
 
 @article{mesri2016optimal,
   author  = {Mesri, Youssef and Khalloufi, M. and Hachem, E.},
-  title   = {On optimal simplicial 3D meshes for minimizing the Hessian-based errors},
+  title   = {On optimal simplicial 3D meshes for minimizing the {Hessian}-based errors},
   journal = {Applied Numerical Mathematics},
   year    = {2016},
   volume  = {109},
@@ -215,5 +231,13 @@
   number  = {18},
   pages   = {2933},
   doi     = {10.3390/math12182933},
+}
+
[…]
