+++ paper.md
@@ -22,19 +22,21 @@ bibliography: paper.bib
 
 Triangular meshes are fundamental data structures in computational science and engineering, serving as the backbone for numerical simulations in fluid dynamics, structural analysis, heat transfer, and many other fields. The quality of a mesh directly impacts the accuracy, stability, and convergence of numerical solutions [@mesri2008dynamic;@manzinali2018adaptive;@Persson2004]. `SOFIA` (Scalable Operators for Field-driven Iso/Ani Adaptation) is a Python library that provides robust and efficient tools for 2D triangular mesh modification, refinement, and quality improvement through local topological operations, with particular emphasis on **anisotropic mesh adaptation** and **automatic boundary preservation**.
 
+In its current version, `SOFIA` focuses on **mesh topology modification (h-adaptation)**. That is, it changes the mesh by locally splitting/collapsing/flipping edges and inserting/removing vertices to refine or coarsen the discretisation. Limited vertex relocation is available through smoothing to improve element quality, but `SOFIA` does not currently implement full **mesh movement (r-adaptation)** driven by a PDE-based moving mesh method; such r-adaptation workflows are left to external solvers or future work.
+
 # Statement of Need
 
 While several mesh generation tools exist [@Shewchuk1996;@Geuzaine2009;@MeshPy;@PyMesh], there is a gap in the Python ecosystem for a lightweight, well-documented library focused specifically on **mesh adaptation** and **local modification operations**. Existing solutions either require complex dependencies (C/C++ bindings), lack comprehensive documentation, or focus primarily on initial mesh generation rather than adaptive remeshing during simulation workflows.
 
-`SOFIA` addresses these needs by providing:
+`SOFIA` addresses these needs by providing a **pure Python** implementation of:
 
-1. **Pure Python implementation with optional C++ acceleration**: Easy to install and modify, with optional C++ backend for performance-critical operations
-2. **Local topological operations**: Fine-grained control over mesh modification through fundamental operations (split, collapse, flip, add, remove)
-3. **Anisotropic mesh adaptation**: Metric-based adaptation with automatic handling of highly stretched elements near features (boundary layers, shocks, fronts)
-4. **Automatic boundary preservation**: Edge collapse that automatically detects and preserves domain boundaries
-5. **Quality-driven adaptation**: Built-in quality metrics and optimization strategies handling both isotropic/anisotropic meshes
-6. **Boundary-aware operations**: Safe manipulation of mesh boundaries with conformity preservation
-7. **Production-ready reliability**: Extensive test suite (100+ unit tests) ensuring robustness
+1. **Local topological operations** (split, collapse, flip, insert, remove) enabling fine-grained mesh modification
+2. **Metric-based anisotropic h-adaptation** driven by a user-supplied tensor field (see below)
+3. **Automatic boundary preservation** built into edge collapse, so boundary conformity can be maintained without manual vertex “locking”
+4. **Quality management** with metrics and optimisation strategies for isotropic and anisotropic meshes
+5. **Reliability** supported by an extensive unit test suite
+
+The repository README mentions a future C++ backend; however, the current release is a pure Python package and this paper describes functionality available in the present version.
 
 The library is designed for computational scientists, researchers, and engineers who need to adaptively refine/coarsen meshes during simulations, optimize existing meshes, implement custom mesh adaptation strategies, or generate boundary layer meshes for high-Reynolds number flows.
 
@@ -48,25 +50,24 @@ The library is designed for computational scientists, researchers, and engineers
[…]
