### 2026-01-29 4a21f3b
Update paper figure to levelset example
