### 2026-01-29 #9729 @youssef-mesri
Thank you for the thoughtful and constructive feedback. We have revised paper.md and paper.bib accordingly. Below we respond point-by-point.
•	GENERAL FEEDBACK
o	Mesh movement literature / r-adaptation vs h-adaptation
We clarified the scope of SOFIA in the paper. The current version focuses on topology modification (h-adaptation) via local operations (split/collapse/flip/insert/remove). We also clarified that SOFIA does not currently implement a full mesh movement (r-adaptation) method; vertex relocation is currently limited to smoothing for mesh quality improvement.
o	Metric-based vs hierarchical h-adaptation
We added clarifying text in the anisotropic adaptation section stating that SOFIA follows a metric-based h-adaptation approach (local mesh modification governed by a user-supplied metric tensor), rather than a hierarchical refinement strategy.
o	Figure request
We added a figure showing an example of an anisotropically adapted mesh produced with SOFIA, included directly in the “Anisotropic Mesh Adaptation” section.
o	Bullet point formatting
We reduced list-heavy formatting where appropriate by converting some list sections into prose (e.g., the “Key Innovation” explanation and the “Key differentiators” section).

•	STATEMENT OF NEED
o	C++ acceleration mention vs repository state
We removed/rewrote the statement to avoid claiming C++ acceleration as currently available functionality. The paper now describes the current pure-Python implementation and mentions the C++ backend only as future work (as stated in the README).

•	ANISOTROPIC MESH ADAPTATION SECTION
o	“up to 10:1 or higher” wording
We removed the redundant phrasing and replaced it with more general wording (“high aspect ratios”).
o	Math notation clarity
We added a short explanation defining the symbols used in the metric edge-length expression (what p1, p2, M, and L_M(e) refer to).

•	STATE OF THE FIELD / COMPARISON SECTION
We revised the “state of the field” discussion to make the mapping between tools and citations explicit and to reduce formatting ambiguity.
o	Citations inside bullet points
We updated the comparison bullets so it is clear which citations correspond to which package/tool (including adding explicit citations for MMG/BAMG in the corresponding bullet)
o	Formatting of “Key differentiators
We corrected the formatting of the “Key differentiators” part (which previously read like an incorrectly formatted bullet list) and rewrote it in prose for improved readability.
o	Firedrake mesh tools reference / PETSc-ParMmg
We added a reference to Wallwork et al. (SIAM IMR 2022, arXiv:2201.02806) and cited it in the Firedrake-related comparison bullet to acknowledge the PETSc/ParMmg metric-adaptation workflow.

•	REFERENCES
o	Consistent capitalization
We harmonized capitalization in paper.bib by protecting key terms with BibTeX braces (e.g., Delaunay, Part I, MATLAB, Hessian, PETSc, ParMmg), ensuring consistent rendering across bibliography styles.
