### 2026-02-04 #9729 @joewallwork
@youssef-mesri Thanks for the updates to the paper, they make it clearer.

Minor points:
* You can probably remove the citations on lines 134-135 given that they are now duplicated in the bullet points.
* I think there is a recommended paper citation for Mmg rather than just the GitHub repo? See https://www.mmgtools.org/publications.html. Perhaps the same is true for FreeFEM (although I'm not familiar with it)?
* The citation for Wallwork et al. 2022 is incorrect - it includes an incorrect author.
* Python should have a capital P in the references.

Other than these minor points, I'm generally happy with the state of the paper and will progress onto the software review.
