### 2026-02-04 e2316b1
paper: fix citations and references
