+++ paper.bib
@@ -87,20 +87,28 @@
   year={2018}
 }
 
-@software{MMG,
-  author    = {{MmgTools contributors}},
-  title     = {{Mmg}: Open source software for simplicial mesh modification},
-  url       = {https://github.com/MmgTools/mmg},
-  publisher = {GitHub},
-  journal   = {GitHub repository}
+@article{MMG,
+  title   = {Three-dimensional adaptive domain remeshing, implicit domain meshing, and applications to free and moving boundary problems},
+  author  = {Dapogny, C. and Dobrzynski, C. and Frey, P.},
+  journal = {Journal of Computational Physics},
+  volume  = {262},
+  pages   = {358--378},
+  year    = {2014},
+  doi     = {10.1016/j.jcp.2014.01.005},
+  url     = {https://www.sciencedirect.com/science/article/pii/S0021999114000266}
 }
 
-@software{BAMG,
-  author    = {{FreeFEM contributors}},
-  title     = {{BAMG}: Bidimensional anisotropic mesh generator/adaptation},
-  url       = {https://github.com/FreeFem/FreeFem-sources},
-  publisher = {GitHub},
-  journal   = {GitHub repository}
+@article{BAMG,
+  AUTHOR = {Hecht, F.},
+  TITLE = {New development in FreeFem++},
+  JOURNAL = {J. Numer. Math.},
+  FJOURNAL = {Journal of Numerical Mathematics},
+  VOLUME = {20}, YEAR = {2012},
+  NUMBER = {3-4}, PAGES = {251--265},
+  ISSN = {1570-2820},
+  MRCLASS = {65Y15},
+  MRNUMBER = {3043640},
+  URL = {https://freefem.org/}
 }
 
 @article{Logg2012,
@@ -235,9 +243,10 @@
 
 @inproceedings{wallwork2022parmmsg,
   title     = {Parallel metric-based mesh adaptation in {PETSc} using {ParMmg}},
-  author    = {Wallwork, James G. and Farrell, Patrick E. and others},
-  booktitle = {SIAM International Meshing Roundtable},
+  author    = {Wallwork, Joseph G. and Knepley, Matthew G. and Barral, Nicolas and Piggott, Matthew D.},
+  booktitle = {International Meshing Roundtable},
   year      = {2022},
+  doi       = {10.48550/arXiv.2201.02806},
   url       = {https://arxiv.org/abs/2201.02806},
 }
 
