### 2026-02-04 #9729 @youssef-mesri
@joewallwork Thanks for the careful read and the helpful suggestions.

- Removed the duplicated citations around lines 134–135 (the bullet points retain the relevant references).
- Updated the MMG and FreeFEM/BAMG references.
- Corrected the Wallwork et al. (2022) citation (author list and details).
- Confirmed “Python” is capitalized consistently.
With these minor fixes in place, we’re happy to proceed to the software review stage, and we look forward to receiving your software review soon.
