### 2026-02-04 1418709
paper: add boundary layer adaptation figure
