### 2026-02-04 #9729 @youssef-mesri
@joewallwork I added another figure to highlight Anisotropic mesh adaptation with boundary layer insertion capability.
