### 2026-02-05 87d417b
paper: add Zenodo DOI reference and availability section
