+++ paper.bib
@@ -250,3 +250,13 @@
   url       = {https://arxiv.org/abs/2201.02806},
 }
 
+@software{sofia2024,
+  author    = {Mesri, Youssef},
+  title     = {{SOFIA}: Scalable Operators for Field-driven Iso/Ani Adaptation},
+  year      = {2024},
+  publisher = {Zenodo},
+  version   = {v0.1.0},
+  doi       = {10.5281/zenodo.18492172},
+  url       = {https://doi.org/10.5281/zenodo.18492172}
+}
+
