### 2026-02-05 37a45a9
docs: add Zenodo DOI badge to README
