### 2026-02-05 #9729 @youssef-mesri
@joewallwork A Zenodo archive has been created for release v0.1.0, and the corresponding DOI has been added to the paper.
