### 2026-02-12 #9729 @hugoledoux
> [@joewallwork](https://github.com/joewallwork) A Zenodo archive has been created for release v0.1.0, and the corresponding DOI has been added to the paper.

@youssef-mesri this is not something you do at this point: focus on the improvement of the code/repo and when it,s finished and accepted you will need to get a DOI and the JOSS team will put it on the paper
