### 2026-02-18 youssef-mesri/sofia-mesh#7 @joewallwork
[JOSS review] Guidance on running demos
Towards https://github.com/openjournals/joss-reviews/issues/9729

I went to run the demos and found that they didn't all work out of the box. For some of them, running from the `demos/` subdirectory works. For others (e.g., `adapt_before_after`) you need to run from the root level (or else there's an error because the default output directory `demos/` doesn't exist. For the `anisotropic_adaptation_naca0012` demo, I needed to edit it to pick up the `meshes/` subdir.

I tried running `python3 -m demos.patch_diagnose`,  `python3 -m demos.patch_batches`, and `python3 -m demos.patch_visual_labels` as recommended in the `demos/README.md` file but these gave the error
```
Traceback (most recent call last):
  File "<frozen runpy>", line 189, in _run_module_as_main
  File "<frozen runpy>", line 112, in _get_module_details
  File "/home/joe/review/joss/sofia/repo/demos/__init__.py", line 22, in <module>
    from .patch_diagnose import run_patch_diagnose
ModuleNotFoundError: No module named 'demos.patch_diagnose'
```

It would be beneficial if the demos README could be updated to give current guidance on how the demos should be run.
