### 2026-02-18 youssef-mesri/sofia-mesh#8 @joewallwork
[JOSS review] Missing `anistropic_remeshing` example
Towards https://github.com/openjournals/joss-reviews/issues/9729

The examples included in SOFIA are great! They cover a nice range of topics, starting from beginner and moving up to advanced. However, there appears to be one missing: 9. `anisotropic_remeshing`. It's mentioned in the `examples/README` but doesn't seem to be included in the repo. There's 12. `anisotropic_remeshing_normalized` but not the other one.
