### 2026-02-18 #9729 @joewallwork
@youssef-mesri I've now had chance to play around with the code. I just wanted to say that the examples provided with SOFIA are great! They proceed nicely from beginner to advanced topics and are well curated in the README. I'm happy to mark the checklist boxes for functionality and examples as complete.

The remaining checklist box that's incomplete is for functionality documentation. As raised in https://github.com/youssef-mesri/sofia-mesh/issues/5, the link to the API docs is broken, so it's unclear whether they exist. Are there API docs for SOFIA? If not, would it be possible to add some?

You will have noticed there are other issues that I've opened directly on the SOFIA repository which I feel it would be good to fix as part of this review.
