### 2026-02-18 #9729 @inducer
A few minor comments from reading the paper:

- Triangle (as wrapped by Meshpy) can do [refinement](https://documen.tician.de/meshpy/tri-tet.html#meshpy.triangle.refine). It might be better to soften the claim regarding being "focused on initial mesh creation".
- "hackable in pure Python" sounds too colloquial. Rephrase?
- Many meshers (including Triangle, TetGen, Gmsh to my knowledge) rely on "robust geometric predicates". (I think the idea goes back to Shewchuk, but I'm not sure.) This requires rather low-level interactions with floating point hardware. Is there a corresponding need in SOFIA? If so, how is it resolved?
- "contributions from the computational geometry and mesh generation communities": This is a bit unspecific. Would it be better to call out specific contributions/contributors?
