### 2026-02-18 youssef-mesri/sofia-mesh#9 @inducer
[JOSS review] `anisotropic_boundary_adaptation.py` fails with "unexpected keyword argument 'use_cpp_core'"
On 37a45a9, running
```
python anisotropic_boundary_adaptation.py
```

I get
```
Traceback (most recent call last):
  File "/home/andreas/tmp/sofia-mesh/examples/anisotropic_boundary_adaptation.py", line 949, in <module>
    sys.exit(main())
             ~~~~^^
  File "/home/andreas/tmp/sofia-mesh/examples/anisotropic_boundary_adaptation.py", line 627, in main
    editor, n_bnd_verts = create_unit_square_mesh(
                          ~~~~~~~~~~~~~~~~~~~~~~~^
        n_boundary=n_boundary,
        ^^^^^^^^^^^^^^^^^^^^^^
        n_interior=n_interior,
        ^^^^^^^^^^^^^^^^^^^^^^
        seed=seed
        ^^^^^^^^^
    )
    ^
  File "/home/andreas/tmp/sofia-mesh/examples/anisotropic_boundary_adaptation.py", line 136, in create_unit_square_mesh
    editor = PatchBasedMeshEditor(all_pts, tri.simplices, use_cpp_core=False)
TypeError: PatchBasedMeshEditor.__init__() got an unexpected keyword argument 'use_cpp_core'
```
cc @joewallwork

x-ref: https://github.com/openjournals/joss-reviews/issues/9729
