### 2026-02-18 youssef-mesri/sofia-mesh#10 @inducer
[JOSS review] Run examples in CI
To prevent issues like #9 in the future, would it make sense to run the examples in CI in an automated fashion?

cc @joewallwork

x-ref https://github.com/openjournals/joss-reviews/issues/9729
