### 2026-02-18 #9729 @inducer
Hi @youssef-mesri, I've just had a chance to work on my review some, and I'm now in a similar spot as @joewallwork I think. The lack of detailed documentation is definitely an issue, and I've filed some more minor things as issues on the repo itself.

I look forward to seeing these stumbling blocks removed, and then I'd be happy to recommend that JOSS proceed with accepting the contribution.
