### 2026-02-19 810d4d5
paper: address reviewer comments - soften Triangle claims, remove geometric predicates discussion, improve acknowledgements
