### 2026-02-19 #9729 @youssef-mesri
Hi @inducer Thank you for your valuable feedback. I have addressed all four points:

Triangle/MeshPy: Softened the claim to acknowledge Triangle's refinement capabilities while clarifying SOFIA's focus on metric-based anisotropic adaptation.

"Hackable": Replaced with "easily extensible for research and prototyping."

Geometric predicates: Removed this discussion entirely as it's not relevant to SOFIA.

Acknowledgements: Made more specific by explicitly thanking JOSS reviewers and mentioning the mesh adaptation research community.

Changes committed in 810d4d5.

Best regards,
Youssef
