### 2026-02-23 c18db79
fix: remove deprecated use_cpp_core parameter
