### 2026-02-23 5f33326
ci: run examples automatically; fix boundary_refinement histogram crash
Co-authored-by: youssef-mesri <60918206+youssef-mesri@users.noreply.github.com>
