+++ .github/workflows/ci.yml
@@ -28,3 +28,28 @@ jobs:
           pip install pytest
       - name: Run tests
         run: pytest -q
+
+  examples:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+          cache: 'pip'
+          cache-dependency-path: |
+            pyproject.toml
+            setup.cfg
+            setup.py
+            requirements.txt
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+      - name: Run examples
+        env:
+          MPLBACKEND: Agg
+        run: bash scripts/test_examples.sh
