### 2026-02-23 c6001cf
ci: add examples job to test all example scripts
