### 2026-02-23 78dd186
docs: remove references to non-existent anisotropic_remeshing.py example
