### 2026-02-23 8f0aa70
docs: fix CONTRIBUTING.md path references in examples README
