### 2026-02-23 3313215
docs: fix version number in README to match tagged release (v0.1.0)
