### 2026-02-23 d305b41
docs: fix API documentation
