+++ docs/API.md
@@ -1,0 +1,530 @@
+# SOFIA API Documentation
+
+Complete reference for SOFIA's mesh modification and analysis tools.
+
+## Table of Contents
+
+- [Core Classes](#core-classes)
+  - [PatchBasedMeshEditor](#patchbasedmesheditor)
+- [Mesh Operations](#mesh-operations)
+  - [Edge Operations](#edge-operations)
+  - [Vertex Operations](#vertex-operations)
+  - [Boundary Operations](#boundary-operations)
+- [Quality and Analysis](#quality-and-analysis)
+- [Remeshing](#remeshing)
+- [I/O Operations](#io-operations)
+- [Utilities](#utilities)
+
+---
+
+## Core Classes
+
+### PatchBasedMeshEditor
+
+The main class for mesh editing operations.
+
+#### Constructor
+
+```python
+from sofia.core.mesh_modifier2 import PatchBasedMeshEditor
+
+editor = PatchBasedMeshEditor(points, triangles)
+```
+
+**Parameters:**
+- `points` (np.ndarray): Nx2 array of vertex coordinates
+- `triangles` (np.ndarray): Mx3 array of triangle vertex indices
+
+**Attributes:**
+- `points` (np.ndarray): Current mesh vertices
+- `triangles` (np.ndarray): Current mesh triangles (inactive triangles marked with -1)
+- `edge_map` (dict): Mapping from edges (sorted tuples) to triangle indices
+- `stats` (dict): Statistics about operations performed
+
+**Example:**
+```python
+import numpy as np
+from scipy.spatial import Delaunay
+
+# Create initial points
+points = np.array([[0, 0], [1, 0], [1, 1], [0, 1]])
+tri = Delaunay(points)
+
+# Create editor
+editor = PatchBasedMeshEditor(points, tri.simplices)
+```
+
+---
+
+## Mesh Operations
+
+### Edge Operations
+
+#### `split_edge(edge, split_point=None, config=None)`
+
+Split an edge, optionally at a specified point.
+
+**Parameters:**
+- `edge` (tuple): Edge as (v1, v2) where v1 < v2
+- `split_point` (np.ndarray, optional): Coordinates for new vertex (default: midpoint)
+- `config` (SplitConfig, optional): Configuration options
+
+**Returns:**
+- `success` (bool): True if operation succeeded
+- `new_vertex` (int or None): Index of newly created vertex
+- `info` (dict): Additional information about the operation
+
+**Example:**
+```python
+# Split edge at midpoint
+success, new_v, info = editor.split_edge((0, 1))
+
+# Split edge at specific point
+split_point = np.array([0.3, 0.5])
+success, new_v, info = editor.split_edge((0, 1), split_point=split_point)
+```
+
+---
+
+#### `edge_collapse(edge, target_point=None, config=None)`
+
+Collapse an edge, merging two vertices into one.
+
+**Parameters:**
+- `edge` (tuple): Edge as (v1, v2) to collapse
+- `target_point` (np.ndarray, optional): Position for merged vertex (default: midpoint)
+- `config` (CollapseConfig, optional): Configuration options
+
+**Returns:**
+- `success` (bool): True if operation succeeded
+- `kept_vertex` (int or None): Index of the vertex that remains
+- `info` (dict): Additional information about the operation
+
+**Example:**
+```python
+# Collapse edge to midpoint
+success, kept_v, info = editor.edge_collapse((0, 1))
+
+# Collapse edge to specific point
+target = np.array([0.2, 0.3])
+success, kept_v, info = editor.edge_collapse((0, 1), target_point=target)
+```
+
+---
+
+#### `flip_edge(edge, config=None)`
+
[…]
