### 2026-02-23 a1742dc
docs: add API documentation link to README and fix paths
