### 2026-02-23 d82e89d
Fix demos to work out of the box
- Remove imports for non-existent modules in demos/__init__.py
- Fix adapt_before_after.py output path (was demos/adapt_pair.png, now adapt_pair.png)
- Add automatic directory creation before saving files
- Update anisotropic_adaptation_naca0012.py print statements
- Update demos/README.md with correct usage instructions

All demos now work when run from project root.
