### 2026-02-23 #9729 @youssef-mesri
> [@youssef-mesri](https://github.com/youssef-mesri) I've now had chance to play around with the code. I just wanted to say that the examples provided with SOFIA are great! They proceed nicely from beginner to advanced topics and are well curated in the README. I'm happy to mark the checklist boxes for functionality and examples as complete.
>
> The remaining checklist box that's incomplete is for functionality documentation. As raised in [youssef-mesri/sofia-mesh#5](https://github.com/youssef-mesri/sofia-mesh/issues/5), the link to the API docs is broken, so it's unclear whether they exist. Are there API docs for SOFIA? If not, would it be possible to add some?
>
> You will have noticed there are other issues that I've opened directly on the SOFIA repository which I feel it would be good to fix as part of this review.

@joewallwork Thank you for the positive feedback on the examples! I'm glad they're helpful and well-organized.

Regarding the API documentation: Yes, comprehensive API docs now exist! I've just added them in commit a1742dc. You can find them here: docs/API.md

The documentation covers:
- PatchBasedMeshEditor class (the main interface)
- Edge, vertex, and boundary operations
- Quality metrics
- Remeshing functions (isotropic and anisotropic)
- I/O operations
- Utility functions

The README has also been updated to link to the API documentation, and the broken link issue (#5) has been resolved.

Regarding the other issues you've opened:
- Demos not working out of the box: Fixed in commit d82e89d. All demos now run successfully from the project root with correct imports and output paths.
- Missing example 9: Documented in the examples README - it was intentionally removed in favor of anisotropic_remeshing_normalized.py which provides better normalization.
- Documentation paths: All fixed - CONTRIBUTING.md is now properly located in docs and all references updated.

All issues you raised have been addressed. Please let me know if there's anything else you'd like to see improved!
