### 2026-02-23 #9729 @youssef-mesri
> I've just had a chance to work on my review some, and I'm now in a similar spot as [@joewallwork](https://github.com/joewallwork) I think. The lack of detailed documentation is definitely an issue, and I've filed some more minor things as issues on the repo itself.
>
> I look forward to seeing these stumbling blocks removed, and then I'd be happy to recommend that JOSS proceed with accepting the contribution.

Thank you for your feedback! All documentation and usability issues have been addressed:

Added:
- Comprehensive API documentation (API.md) covering all major classes and functions
- Fixed all broken documentation links
- Automated CI testing for all examples
- Resolved Issues:
 Demos now work out of the box (fixed imports and paths)
 Documentation complete and properly linked
 All example scripts tested and working

All changes are now live on the main branch. Please let me know if anything else needs attention!
