### 2026-02-23 youssef-mesri/sofia-mesh#7 @joewallwork
Thanks, this does indeed satisfy my requests.
