### 2026-02-23 youssef-mesri/sofia-mesh#8 @joewallwork
I'm not sure this issue has been resolved? The original `anisotropic_remeshing` example is still mentioned in the README.
