### 2026-02-23 #9729 @joewallwork
Thanks for making all of these changes @youssef-mesri.

I was interested that you added the API documentation manually. For the sake of reducing the maintenance burden, it might be a good idea to investigate setting up an automated solution for this going forward. (Not necessarily as part of this review, though.) It can be done in Python with [Sphinx](https://www.sphinx-doc.org), for example.

Regarding the issue I raised about the missing `anisotropic_remeshing` example, I'm not sure you've actually resolved the issue? The example is still mentioned in the examples README.
