### 2026-02-24 3386597
docs: fix paths
