### 2026-02-24 3a7458e
docs: fix examples README - remove references to non-existent files and duplicated content
