+++ examples/README.md
@@ -24,17 +24,7 @@ Learn the fundamental mesh operations:
 - Flip edges to improve quality
 - Collapse edges to simplify
 - Check mesh quality metrics
-- VisuaComplete Workflow (15 min)
-
-**Track 4 - Anisotropic Adaptation (Advanced to Expert):**
-11. Anisotropic Level-Set Adaptation → 
-12. Anisotropic Remeshing with Normalization [Expert] → 10. Anisotropic Boundary Adaptation
-         
-Anisotropic Level-Set Adaptation (20 min)
-         
-Anisotropic Remeshing with Normalization (30 min) [Expert]
-         
-Anisotropic Boundary Adaptation (15 min)-by-step results
+- Visualize step-by-step results
 
 ```bash
 python examples/basic_remeshing.py
@@ -221,48 +211,7 @@ python examples/combined_refinement.py
 
 ---
 
-### 9. Anisotropic Remeshing (`anisotropic_remeshing.py`)
-**Difficulty:** Advanced  
-**Time:** ~15 minutes
-
-Metric-guided anisotropic mesh adaptation:
-- Build random Delaunay mesh
-- Define spatially-varying metric tensor field M(x)
-- Compute metric edge lengths L_M(e) = sqrt((q-p)^T M (q-p))
-- Iteratively refine/coarsen based on metric lengths
-- Split edges where L_M > alpha, collapse where L_M < beta
-- Maintain mesh conformity throughout
-- Visualize metric ellipses showing directional scaling
-
-```bash
-python examples/anisotropic_remeshing.py
-```
-
-**Output:** `anisotropic_remeshing_result.png` (6-panel comprehensive visualization)
-
-**Technical Highlights:**
-- Metric tensor M(x): 2×2 symmetric positive-definite matrix at each point
-- Pure Python implementation using split/collapse operations only
-- Preserves mesh conformity (each edge shared by ≤2 triangles)
-- Blue ellipses visualize the anisotropic field (axes = eigenvectors, scaling = eigenvalues)
-
-**Results Example:**
-- Initial: 84 vertices, 162 triangles
-- After: 23 vertices, 37 triangles (-77% triangles, significant coarsening)
-- Operations: 3 splits, 64 collapses, 0 flips
-- Conformity: Maintained throughout
-- Metric deviation: 0.996 -> 0.898 (9.8% improvement)
-
-**Perfect for:** Understanding metric-guided adaptation, anisotropic features
-
-**Key Concepts:** Metric tensors, Riemannian geometry, anisotropic adaptation, conformity preservation
-
-**Note:** This example uses split/collapse only to maintain conformity in pure Python.
-For full remeshing with flips and metric-space smoothing, see `demos/adapt_scenario.py` with C++ core.
-
----
-
-### 11. Anisotropic Level-Set Adaptation (`anisotropic_levelset_adaptation.py`)
+### 9. Anisotropic Level-Set Adaptation (`anisotropic_levelset_adaptation.py`)
 **Difficulty:** Advanced  
 **Time:** ~20 minutes
 
@@ -306,7 +255,7 @@ python examples/anisotropic_levelset_adaptation.py
 
 ---
 
-### 12. Anisotropic Remeshing with Metric Normalization (`anisotropic_remeshing_normalized.py`)
+### 10. Anisotropic Remeshing with Metric Normalization (`anisotropic_remeshing_normalized.py`)
 **Difficulty:** Expert  
 **Time:** ~30 minutes
 
@@ -372,16 +321,12 @@ python examples/anisotropic_remeshing_normalized.py --no-normalize
 - Empirical calibration for remeshing algorithms
 - Mesh quadrature for efficient integration
 
-**Documentation:**
-- See `README_metric_normalization.md` for theory and detailed usage
[…]
