### 2026-02-24 c6a4fb6
docs: remove obsolete EXAMPLES_OVERVIEW.md
