+++ examples/EXAMPLES_OVERVIEW.md
@@ -1,330 +1,0 @@
-# Sofia Examples Overview
-
-This directory contains example scripts demonstrating various mesh operations and remeshing techniques.
-
-## Quick Reference: Anisotropic Remeshing Examples
-
-| Example | Difficulty | Time | Key Feature | When to Use |
-|---------|-----------|------|-------------|-------------|
-| `anisotropic_remeshing.py` | Advanced | 15 min | Basic metric-guided adaptation | Learn fundamentals of anisotropic remeshing |
-| `anisotropic_levelset_adaptation.py` | Advanced | 20 min | Level-set based metric | Adapt to curved features and iso-contours |
-| `anisotropic_remeshing_normalized.py` | Expert | 30 min | Metric normalization + budget control | Target specific vertex count, production use |
-| `anisotropic_boundary_adaptation.py` | Advanced | 15 min | Boundary layer insertion | CFD preprocessing, boundary layer meshing |
-
-**Recommended learning path:** 
-9 → 11 → 12 (anisotropic_remeshing → levelset → normalized)
-
----
-
-## Main Examples
-
-### Anisotropic Remeshing
-
-#### `anisotropic_levelset_adaptation.py` **Recommended**
-A clean, well-documented implementation of anisotropic remeshing with level-set based metric and boundary preservation.
-
-**Features:**
-- Anisotropic metric field based on distance to a level-set function (sinusoidal curve)
-- Automatic boundary preservation (no manual protection needed for collapse)
-- Split/collapse/smooth operations
-- Comprehensive visualization with statistics
-- Quality checks disabled for anisotropic meshes (elongated triangles are desired)
-
-**Usage:**
-```bash
-python anisotropic_levelset_adaptation.py --max-iter 10 --alpha 0.8 --beta 0.7
-```
-
-**Key innovations:**
-- Level-set function φ(x,y) = y - (0.5 + 0.15·sin(4πx)) defines the feature
-- Metric adapts based on distance to iso-contours
-- Boundary vertices automatically detected and preserved during edge collapse
-- Quality checks based on angles are disabled (not suitable for anisotropic meshes)
-- C++ backend disabled to use Python implementation with boundary detection
-
-**Results:**
-- Error reduction: ~3x
-- Metric edge lengths: well-aligned to ideal (L_M ≈ 1.0)
-- Boundary perfectly preserved (deviation = 0.00)
-
-#### `anisotropic_remeshing_normalized.py` **Expert-Level**
-Advanced anisotropic remeshing with **metric normalization** for precise computational budget control.
-
-**Features:**
-- All features of `anisotropic_levelset_adaptation.py` plus:
-- **Metric normalization**: Target specific vertex count via determinant normalization
-- **Mesh quadrature integration**: 50× faster than grid sampling (0.01s vs 0.5s)
-- **Independent control**: Decouple anisotropy ratio from vertex count
-- Calibration factor for accurate vertex count targeting
-
-**Usage:**
-```bash
-# Target 350 vertices
-python anisotropic_remeshing_normalized.py --target-complexity 350 --alpha 1.3 --beta 0.5
-
-# Target 500 vertices with more iterations
-python anisotropic_remeshing_normalized.py --target-complexity 500 --max-iter 10
-
-# Disable normalization (original behavior)
-python anisotropic_remeshing_normalized.py --no-normalize
-```
-
-**Key innovation - Metric Normalization:**
-- Formula: M_normalized = α² · M where α = sqrt(N_target / (C · ∫√det(M) dΩ))
[…]
