### 2026-02-24 53c5e29
docs: fix examples README - remove references to C++ backend
