### 2026-02-24 #9729 @youssef-mesri
@joewallwork
You’re right about the API documentation - automating it would reduce maintenance burden.

Regarding anisotropic_remeshing, I’ve now fixed the README reference. Could you please check that everything looks correct on your side?

Many thanks!
