### 2026-02-24 #9729 @joewallwork
> Regarding anisotropic_remeshing, I’ve now fixed the README reference. Could you please check that everything looks correct on your side?

Thanks for fixing that @youssef-mesri.

@hugoledoux: The author has addressed all of my requests and my reviewer checklist is complete so I'm happy to recommend the paper for publication.
