### 2026-02-25 #9729 @inducer
As with @joewallwork, I'm now happy with how things look (and my checklist is complete), so I am happy to recommend publication as well.
