### 2026-03-02 #9729 @youssef-mesri
@joewallwork and @inducer thank you for the careful evaluation and for the constructive feedback provided throughout the review process.
@hugoledoux I'm very pleased to learn that the reviewers have indicated that all requests have been addressed and have recommended the paper for publication. In light of this positive recommendation, I kindly ask whether the manuscript may now proceed to the final stage of acceptance and publication. Please let me know if any additional information or formalities are required from my side.

Thanks for your time and consideration!
