### 2026-04-17 #9729 @hugoledoux
I am taking care of this today, sorry for delay (and holidays in the middle)
