### 2026-04-21 #9729 @hugoledoux
> [@youssef-mesri](https://github.com/youssef-mesri) as someone who has implemented triangulators in 2D and 3D, I am a bit surprised that you mentioned above that robust predicates are "not relevant to SOFIA" (while I couldn't make my triangulators work for all cases without using Shewchuk's work). Can you elaborate on this please, I am curious.

@youssef-mesri I think you overlooked this?
