### 2026-04-21 #9729 @youssef-mesri
> > [@youssef-mesri](https://github.com/youssef-mesri) as someone who has implemented triangulators in 2D and 3D, I am a bit surprised that you mentioned above that robust predicates are "not relevant to SOFIA" (while I couldn't make my triangulators work for all cases without using Shewchuk's work). Can you elaborate on this please, I am curious.
>
> [@youssef-mesri](https://github.com/youssef-mesri) I think you overlooked this?

@hugoledoux thanks for the reminder. I may not have stated this explicitly in the paper text itself, but SOFIA's design does not rely on robust predicates. To clarify: SOFIA is not a triangulator and does not construct a Delaunay triangulation from scratch. It locally modifies an already valid mesh through split, collapse, flip, and vertex-insertion/remove operations. In this context, the standard motivation for robust geometric predicates (orientation and in-circle tests needed to maintain a global Delaunay property during insertion) does not apply.
SOFIA's robustness during local cavity re-triangulation is ensured by a different mechanism: only re-triangulations that exactly conserve the total area of the modified cavity are accepted. This area-preservation check filters out degenerate or inverted configurations without relying on exact predicate evaluation.
Anisotropic Delaunay-based methods (which do require robust predicates to enforce the anisotropic empty-circumball property) follow a different algorithmic path. SOFIA's local-modification approach is described in Mesri et al. (2008, IMR) and Mesri et al. (2012, AMC).
