### 2026-04-28 #9729 @hugoledoux
thanks for explanation, this is clearer (and I agree not necessary in the paper).
