### 2026-05-04 dbecf8e
bump version to 1.0.0 for JOSS/Zenodo release
