+++ paper.bib
@@ -255,7 +255,7 @@
   title     = {{SOFIA}: Scalable Operators for Field-driven Iso/Ani Adaptation},
   year      = {2024},
   publisher = {Zenodo},
-  version   = {v0.1.0},
+  version   = {v1.0.0},
   doi       = {10.5281/zenodo.18492172},
   url       = {https://doi.org/10.5281/zenodo.18492172}
 }
