+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = sofia-mesh
-version = 0.1.0
+version = 1.0.0
 
 [options]
 package_dir =
