+++ paper.bib
@@ -256,7 +256,7 @@
   year      = {2024},
   publisher = {Zenodo},
   version   = {v1.0.0},
-  doi       = {10.5281/zenodo.18492172},
-  url       = {https://doi.org/10.5281/zenodo.18492172}
+  doi       = {10.5281/zenodo.20053887},
+  url       = {https://doi.org/10.5281/zenodo.20053887}
 }
 
