+++ paper.bib
@@ -71,7 +71,7 @@
 
 @software{MeshPy,
   author={Kloeckner, Andreas},
-  title={MeshPy: Simplicial Mesh Generation from Python},
+  title={MeshPy: Simplicial Mesh Generation from {P}ython},
   url={https://mathema.tician.de/software/meshpy/},
   publisher = {Mathema},
   journal = {Website repository},
@@ -80,7 +80,7 @@
 
 @software{PyMesh,
   author={Zhou, Qingnan},
-  title={PyMesh: Geometry Processing Library for Python},
+  title={PyMesh: Geometry Processing Library for {P}ython},
   url={https://github.com/PyMesh/PyMesh},
   publisher = {GitHub},
   journal = {GitHub repository},
