+++ paper.bib
@@ -108,6 +108,7 @@
   ISSN = {1570-2820},
   MRCLASS = {65Y15},
   MRNUMBER = {3043640},
+  DOI = {10.1515/jnum-2012-0013},
   URL = {https://freefem.org/}
 }
 
