+++ paper.bib
@@ -180,7 +180,6 @@
   doi={10.1137/090754078}
 }
 
-% --- Added Mesri (Y.) publications collected from Google Scholar / publisher pages ---
 @article{mesri2006continuous,
   title={Continuous mesh adaptation models for CFD},
   author={Mesri, Youssef and Alauzet, Fr{\'e}d{\'e}ric and Loseille, Adrien and Hasco{\"e}t, Laurent and Koobus, Bruno and Dervieux, Alain},
